requirements.txt
transformers>=4.44.0
datasets
accelerate
peft
bitsandbytes; platform_system == "Linux"
openai>=1.40.0
supabase>=2.4.0
python-dotenv
tiktoken
ujson
pydantic>=2